requirements.txt
alembic==0.8.10
Babel==2.3.4
bcrypt==3.1.2
blinker==1.4
boto3==1.4.4
botocore==1.5.1
cffi==1.9.1
click==6.7
docutils==0.13.1
Flask==0.12
Flask-Assets==0.12
Flask-Babel==0.11.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.4
jmespath==0.9.0
Mako==1.0.6
MarkupSafe==0.23
mysqlclient==1.3.9
passlib==1.7.0
pycparser==2.17
python-dateutil==2.6.0
python-editor==1.0.3
python-magic==0.4.12
pytz==2016.10
s3transfer==0.1.10
six==1.10.0
SQLAlchemy==1.1.5
SQLAlchemy-Utils==0.32.12
webassets==0.12.1
Werkzeug==0.11.15
WTForms==2.1
WTForms-JSON==0.3.0
